Fixed a compilation error due to type ambiguity (ambiguous overload)

+++ src/NTPClientLib.cpp
@@ -221,7 +221,7 @@ void NTPClient::dnsFound (const ip_addr_t *ipaddr) {
     responseTimer2.detach ();
     ntpServerIPAddress = getIPClass (ipaddr);
     DEBUGLOG ("%s - %s\n", __FUNCTION__, ntpServerIPAddress.toString ().c_str ());
-    if (ipaddr != NULL && ntpServerIPAddress != 0)
+    if (ipaddr != NULL && ntpServerIPAddress != (uint32_t)(0))
       setTime (getTime ());
 }
